3Dモデルにターゲットポインターを追加
